Moved the rest packages into the rest subdir to mimic the models setup.
