+++ rest/authentication/util.go
