+++ rest/sessions/util.go
